+++ packages/core/src/formatters/array/set.formatter.ts
@@ -1,7 +1,7 @@
 import { Utils } from '../../services/utils';
 
 /**
- * Set property of object, array or value
+ * Sets property of object, array or value
  * @see https://gist.github.com/der-On/cdafe908847e2b882691
  */
 export const setFormatter = {