+++ packages/core/src/formatters/index.ts
@@ -1,5 +1,6 @@
+export * from './array/index';
 export * from './compare/index';
 export * from './math/index';
-export * from './property/index';
 export * from './special/index';
 export * from './string/index';
+export * from './type/index';