+++ packages/core/src/formatters/special/boolean.formatter.ts
@@ -1,9 +1,0 @@
-/**
- * Parse a string to boolean
- */
-export const booleanFormatter = {
-  name: 'boolean',
-  read(value: string | boolean) {
-    return value === 'true' || value === true;
-  },
-};