+++ packages/core/src/formatters/special/index.ts
@@ -1,8 +1,5 @@
 // special helper formatters
 export { argsFormatter } from './args.formatter';
-export { booleanFormatter } from './boolean.formatter';
 export { debugFormatter } from './debug.formatter';
 export { defaultFormatter } from './default.formatter';
 export { callFormatter } from './call.formatter';
-export { currencyFormatter } from './currency.formatter';
-export { jsonFormatter } from './json.formatter';