+++ packages/core/src/formatters/string/match.formatter.ts
@@ -1,6 +1,9 @@
 export const matchFormatter = {
   name: 'match',
   read(a: string, regexp: string, flags?: string) {
+    if (!a || !regexp) {
+      return false;
+    }
     return a.match(new RegExp(regexp, flags));
   },
 };