+++ packages/core/src/formatters/string/replace-first.formatter.ts
@@ -5,7 +5,7 @@
  * @see https://help.shopify.com/en/themes/liquid/filters/string-filters#replace_first
  */
 export const replaceFirstFormatter = {
-  name: 'replace_first',
+  name: 'replaceFirst',
   read(str: string, value: string, replaceValue: string) {
     return str.replace(value, replaceValue);
   },