+++ packages/core/src/formatters/type/is-boolean.formatter.ts
@@ -1,0 +1,10 @@
+import { IFormatter } from '../../interfaces/formatter';
+import { Utils } from '../../services/utils';
+
+/**
+ * Checks if value is an boolean
+ */
+export const isBooleanFormatter: IFormatter = {
+  name: 'isBoolean',
+  read: Utils.isBoolean,
+};