+++ packages/core/src/formatters/type/is-defined.formatter.ts
@@ -2,7 +2,7 @@ import { IFormatter } from '../../interfaces/formatter';
 import { Utils } from '../../services/utils';
 
 /**
- * Check if value is defined
+ * Checks if value is defined
  */
 export const isDefinedFormatter: IFormatter = {
   name: 'isDefined',