+++ packages/core/src/formatters/type/is-number.formatter.ts
@@ -1,0 +1,10 @@
+import { IFormatter } from '../../interfaces/formatter';
+import { Utils } from '../../services/utils';
+
+/**
+ * Checks if value is a number
+ */
+export const isNumberFormatter: IFormatter = {
+  name: 'isNumber',
+  read: Utils.isNumber,
+};