+++ packages/core/src/formatters/type/is-object.formatter.ts
@@ -2,7 +2,7 @@ import { IFormatter } from '../../interfaces/formatter';
 import { Utils } from '../../services/utils';
 
 /**
- * Check if value is undefined
+ * Checks if value is undefined
  */
 export const isObjectFormatter: IFormatter = {
   name: 'isObject',