+++ packages/bs4/src/components/index.ts
@@ -1,6 +1,8 @@
-export { Bs4ScrollspyComponent } from './bs4-scrollspy/bs4-scrollspy.component';
-export { TabsComponent } from './tabs/tabs.component';
-export { DropdownComponent } from './dropdown/dropdown.component';
-export { NavbarComponent } from './navbar/navbar.component';
 export { Bs4ContentsComponent } from './bs4-contents/bs4-contents.component';
 export { Bs4IconComponent } from './bs4-icon/bs4-icon.component';
+export { Bs4ScrollspyComponent } from './bs4-scrollspy/bs4-scrollspy.component';
+export { Bs4SidebarComponent } from './bs4-sidebar/bs4-sidebar.component';
+export { Bs4ToggleButtonComponent } from './bs4-toggle-button/bs4-toggle-button.component';
+export { DropdownComponent } from './dropdown/dropdown.component';
+export { NavbarComponent } from './navbar/navbar.component';
+export { TabsComponent } from './tabs/tabs.component';