+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Dev
+## 1.4.0
 
 ### Core
 - Add support for attributes with (normally invalid) single qoute json strings
@@ -12,6 +12,9 @@
 ### BS4
 - New Bs4SidebarComponent
 - New Bs4ToggleButtonComponent
+
+### Router
+- New option to not update the browser url
 
 ## 1.3.0